Update optimizers.py
Modified the init_iterate_fn and forward function of BaseOptim class to make them compatible with weight matrix  and value L. Also modified the create_iterator function for the same reason. The variable names alpha_1 and alpha_2 are hard-coded, so we will run into errors if the names are changed or we are using a different architecture. Additionally, create_iterator function has also been modified to ensure that lambda is always positive.