Update unfolded.py
Modified the forward function of BaseUnfold to make it compatible with weight matrix W and value L for distinct step sizes. The variable names alpha_1 and alpha_2 are hard-coded, so we will run into errors if the names are changed or we are using a different architecture. Additionally, new_lambda is used for computing input-specific lambda. If that is not required, comment out those lines and uncomment new_lambda=None.